requirements.txt
pandas==1.3.5
Pillow
python-dotenv
quickchart.io
yfinance
